requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
